fix: fix copying of svg assets, move storybook assets
